Support game screens!
A Game Screen consists of a collection of graphical and logical elements that we want to treat as a single unit, e.g. a menu that we want to display, allow the player to interact with, and then remove from the window. We'll use the SnakeGame as a thin wrapper that manages forwarding KeyEvents to the active screen and changing between screens. A given game screen internally manages their state, what to display and when, and of notifying that the screen should change.